Fix unsaved modal and move routes to separate file
