+++ frontend/schemaCMS/src/shared/components/addBlockForm/addBlockForm.component.js
@@ -121,9 +121,11 @@ export const AddBlockForm = ({ fetchBlockTemplates, projectId, backUrl, blockTem
     if (!customLocation && !goToLocation.pathname !== match.url) {
       setCustomLocation(goToLocation);
       setLeavingPageModalOpen(true);
+
+      return false;
     }
 
-    return false;
+    return true;
   };
 
   const handleBackButtonClick = () => {
